+++ renderer-core/src/main/java/top/ceroxe/rt/renderer/rt/pipeline/RtCommandBufferUploads.java
@@ -45,7 +45,7 @@ final class RtCommandBufferUploads {
                 );
             }
         } finally {
-            MemoryUtil.memFree(upload);
+            MemoryUtil.nmemFree(MemoryUtil.memAddress0(upload));
         }
     }