+++ renderer-core/src/main/java/top/ceroxe/rt/renderer/rt/pipeline/RtRayTracingPipeline.java
@@ -1639,7 +1639,7 @@ public final class RtRayTracingPipeline implements AutoCloseable {
                 }
             }
         } finally {
-            MemoryUtil.memFree(upload);
+            MemoryUtil.nmemFree(MemoryUtil.memAddress0(upload));
         }
     }